openhome/requirements.txt
sounddevice~=0.4.6
soundfile~=0.12.1
numpy~=1.26.3
openai==0.28.0
colorama~=0.4.6
pydub~=0.25.1
pyyaml~=6.0.1
SpeechRecognition~=3.10.1
fuzzywuzzy~=0.18.0
python-Levenshtein~=0.24.0
spotipy~=2.23.0